issue #1200 : Alternative for system property settings
Alternative for system property settings that doesn't set a single value for a setting for a system lifetime. Alternatively relies on the pre-existing RioConfig.useDefaults to control a cached set of system property settings.

The new convert method on RioSetting has a safe default implementation, so does not impact backwards compatibility with other RioSetting implementations.

This is only a stub implementation for review. A full implementation would have new subclasses of RioSettingImpl for each type that we support.

Also does not have an implementation of RioConfigurationException so does not currently compile

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>